Moved test for #1647 regression to regressions file.
